Refactor dbus linked devices interface
Export a separate dbus object for each device

+++ lib/src/main/java/org/asamk/signal/manager/Manager.java
@@ -117,7 +117,7 @@ public interface Manager extends Closeable {
 
     List<Device> getLinkedDevices() throws IOException;
 
-    void removeLinkedDevices(int deviceId) throws IOException;
+    void removeLinkedDevices(long deviceId) throws IOException;
 
     void addDeviceLink(URI linkUri) throws IOException, InvalidKeyException;
